Move NonEmptyListSyntax to org.http4s.syntax and streamline other syntax
