+++ core/src/main/scala/org/http4s/headers/Accept-Charset.scala
@@ -3,7 +3,7 @@ package headers
 
 import cats.data.NonEmptyList
 import org.http4s.parser.HttpHeaderParser
-import org.http4s.util.nonEmptyList._
+import org.http4s.syntax.nonEmptyList._
 
 object `Accept-Charset` extends HeaderKey.Internal[`Accept-Charset`] with HeaderKey.Recurring {
   override def parse(s: String): ParseResult[`Accept-Charset`] =
