+++ core/src/main/scala/org/http4s/headers/Connection.scala
@@ -3,9 +3,8 @@ package headers
 
 import cats.data.NonEmptyList
 import org.http4s.parser.HttpHeaderParser
-import org.http4s.syntax.string._
+import org.http4s.syntax.all._
 import org.http4s.util.{CaseInsensitiveString, Writer}
-import org.http4s.util.nonEmptyList._
 
 // values should be case insensitive
 //http://stackoverflow.com/questions/10953635/are-the-http-connection-header-values-case-sensitive
