+++ core/src/main/scala/org/http4s/syntax/AllSyntax.scala
@@ -7,5 +7,6 @@ trait AllSyntax
     with EffectResponseSyntax
     with EffectRequestSyntax
     with KleisliSyntax
+    with NonEmptyListSyntax
     with StringSyntax
     with LiteralsSyntax
