+++ core/src/main/scala/org/http4s/syntax/LiteralsSyntax.scala
@@ -4,7 +4,8 @@ package syntax
 import org.http4s.internal.Macros
 
 trait LiteralsSyntax {
-  implicit def http4sLiteralsSyntax(sc: StringContext) = new LiteralsOps(sc)
+  implicit def http4sLiteralsSyntax(sc: StringContext): LiteralsOps =
+    new LiteralsOps(sc)
 }
 
 class LiteralsOps(val sc: StringContext) extends AnyVal {
